aider: Corrected the regular expression syntax error by properly escaping the backslashes in `check_generic_api_key` function.
# Aider chat conversation:

USER: SyntaxError: unexpected character after line continuation character
/Users/djh/.cache/pre-commit/repoo8gme5uy/py_env-python3.12/lib/python3.12/site-packages/hooks/forbid_secrets.py:38: SyntaxWarning: invalid escape sequence '\s'
  return re.search(r'(?i)api(_|-)?key['"\s]?[:=]['"\s]?[0-9a-zA-Z]{32,}', content)
Traceback (most recent call last):
  File "/Users/djh/.cache/pre-commit/repoo8gme5uy/py_env-python3.12/bin/forbid_secrets", line 5, in <module>
    from hooks.forbid_secrets import main
  File "/Users/djh/.cache/pre-commit/repoo8gme5uy/py_env-python3.12/lib/python3.12/site-packages/hooks/forbid_secrets.py", line 38
    return re.search(r'(?i)api(_|-)?key['"\s]?[:=]['"\s]?[0-9a-zA-Z]{32,}', content)

ASSISTANT: It appears there is a syntax error in the regular expression within the `forbid_secrets.py` file. The error is due to an invalid escape sequence. I will correct this by properly escaping the backslashes in the regular expression.

Here is the diff for the necessary changes:

```diff
--- hooks/forbid_secrets.py
+++ hooks/forbid_secrets.py
@@ ... @@
-def check_generic_api_key(content):
-    return re.search(r'(?i)api(_|-)?key['"\s]?[:=]['"\s]?[0-9a-zA-Z]{32,}', content)
+def check_generic_api_key(content):
+    return re.search(r'(?i)api(_|-)?key[\'"\\s]?[:=][\'"\\s]?[0-9a-zA-Z]{32,}', content)
```

This change corrects the regular expression by escaping the backslash that precedes the 's' character, which is intended to match whitespace characters.