aider: Specify local path for sops-pre-commit plugin in .pre-commit-config.yaml.
# Aider chat conversation:

USER: If this plugin is stored at /Users/djh/work/src/github.com/djh00t/sops-pre-commit/ how do I make my .pre-commit-config.yaml use it locally?

ASSISTANT: To use the plugin locally, you need to modify the `.pre-commit-config.yaml` file to point to the local path of the plugin instead of the remote repository URL. Here's the diff that shows the changes you need to make:

```diff
--- .pre-commit-config.yaml
+++ .pre-commit-config.yaml
@@ ... @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
   rev: v4.3.0
   hooks:
   - id: check-added-large-files
   - id: check-merge-conflict
   - id: detect-private-key
   - id: trailing-whitespace
-- repo: https://github.com/djh00t/sops-pre-commit
+- repo: local
   rev: v1.0.0
   hooks:
+    - id: kubernetes-secret
+      name: Check for unencrypted Kubernetes secrets
+      entry: python /Users/djh/work/src/github.com/djh00t/sops-pre-commit/hooks/forbid_secrets.py --hook-id kubernetes-secret
+      language: system
+      files: ((^|/)*.(ya?ml)$)
+    # Add similar entries for other hooks as needed
   # ... (other repositories and hooks)
```

This change specifies that the repository is local and provides the path to the `forbid_secrets.py` script for the `kubernetes-secret` hook. You should add similar entries for other hooks as needed, following the same pattern.